+++ src/components/Breadcrumbs/CurrentBreadcrumbItem.tsx
@@ -1,4 +1,4 @@
-import { Badge, BadgeProps } from "@components/Badge/Badge";
+import { Badge, BadgeProps } from "@components/Badge";
 import { useBreadcrumbItem } from "@react-aria/breadcrumbs";
 import { useFocusRing } from "@react-aria/focus";
 import { mergeProps } from "@react-aria/utils";
